fix: move JavaScalarsWiringPostProcessor to graphql-jpa-query-scalars
