+++ graphql-jpa-query-boot-starter-graphql/src/main/java/com/introproventures/graphql/jpa/query/boot/autoconfigure/JavaScalarsRuntimeWiringConfigurer.java
@@ -1,7 +1,7 @@
 package com.introproventures.graphql.jpa.query.boot.autoconfigure;
 
-import com.introproventures.graphql.jpa.query.autoconfigure.JavaScalarsWiringPostProcessor;
 import com.introproventures.graphql.jpa.query.schema.JavaScalars;
+import com.introproventures.graphql.jpa.query.schema.JavaScalarsWiringPostProcessor;
 import graphql.schema.idl.RuntimeWiring;
 import org.springframework.graphql.execution.RuntimeWiringConfigurer;
